🍊 implement endpoints for records and audit the entities

+++ src/main/java/xabufas/rubiks_rush_be/models/Record.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class Record {
-    private float time;
+    private Double time;
     private String scramble;
 }
